Convert references to reST citations
Convert the original reference section, with hard-coded square bracket
reference tokens, to native reST citations. Depending on the type of
reference, I actually re-formattted the references into three varieties:

- One-off lines to web pages are presented as inline links. This seems
  more web-native Some of the original
- Links to product home pages (e.g., to a DB) as coded as reference
  targets so that, e.g. `Name`_ is linked.
- Real references are converted to the reST citation syntax, where the
  reference tokens are named.

I'm not entirely happy with the consistency of reference handling. This
exercise shows that it will be good to port latex's natbib over to reST.

For DM-3835.